+++ debugsnapshot.c
@@ -8,9 +8,7 @@
 #include <np2ver.h>
 #include <statsave.h>
 
-#if defined(__LIBRETRO__)
-#include <rhash.h>
-#elif(_MSC_VER)
+#if(_MSC_VER)
 #include <windows.h>
 #else
 #include <openssl/sha.h>
@@ -26,6 +24,332 @@
 UINT NP2_DebugSnapshot_Count;
 
 NP2_DebugSnapshot_t tDebugSnapshot;
+
+#if defined(__LIBRETRO__)
+static int sha1_calculate(const char *path, char *result);
+
+struct sha1_context
+{
+   unsigned Message_Digest[5]; /* Message Digest (output)          */
+
+   unsigned Length_Low;        /* Message length in bits           */
+   unsigned Length_High;       /* Message length in bits           */
+
+   unsigned char Message_Block[64]; /* 512-bit message blocks      */
+   int Message_Block_Index;    /* Index into message block array   */
+
+   int Computed;               /* Is the digest computed?          */
+   int Corrupted;              /* Is the message digest corruped?  */
+};
+
+static void SHA1Reset(struct sha1_context *context);
+static int SHA1Result(struct sha1_context *context);
+static void SHA1Input(struct sha1_context *context,
+      const unsigned char *message_array,
+      unsigned length);
+
+/* SHA-1 implementation. */
+
+/*
+ *  sha1.c
+ *
+ *  Copyright (C) 1998, 2009
+ *  Paul E. Jones <paulej@packetizer.com>
+ *  All Rights Reserved
+ *
+ *****************************************************************************
+ *  $Id: sha1.c 12 2009-06-22 19:34:25Z paulej $
+ *****************************************************************************
+ *
+ *  Description:
+ *      This file implements the Secure Hashing Standard as defined
+ *      in FIPS PUB 180-1 published April 17, 1995.
+ *
+ *      The Secure Hashing Standard, which uses the Secure Hashing
+ *      Algorithm (SHA), produces a 160-bit message digest for a
+ *      given data stream.  In theory, it is highly improbable that
+ *      two messages will produce the same message digest.  Therefore,
+ *      this algorithm can serve as a means of providing a "fingerprint"
+ *      for a message.
+ *
+ *  Portability Issues:
+ *      SHA-1 is defined in terms of 32-bit "words".  This code was
+ *      written with the expectation that the processor has at least
+ *      a 32-bit machine word size.  If the machine word size is larger,
+ *      the code should still function properly.  One caveat to that
+ *      is that the input functions taking characters and character
+ *      arrays assume that only 8 bits of information are stored in each
+ *      character.
+ *
+ *  Caveats:
+ *      SHA-1 is designed to work with messages less than 2^64 bits
+ *      long. Although SHA-1 allows a message digest to be generated for
+ *      messages of any number of bits less than 2^64, this
+ *      implementation only works with messages with a length that is a
+ *      multiple of the size of an 8-bit character.
+ *
+ */
+
+/* Define the circular shift macro */
+#define SHA1CircularShift(bits,word) ((((word) << (bits)) & 0xFFFFFFFF) | ((word) >> (32-(bits))))
+#if 0
+struct sha1_context
+{
+   unsigned Message_Digest[5]; /* Message Digest (output)          */
+
+   unsigned Length_Low;        /* Message length in bits           */
+   unsigned Length_High;       /* Message length in bits           */
+
+   unsigned char Message_Block[64]; /* 512-bit message blocks      */
+   int Message_Block_Index;    /* Index into message block array   */
+
+   int Computed;               /* Is the digest computed?          */
+   int Corrupted;              /* Is the message digest corruped?  */
+};
+#endif
+
+static void SHA1Reset(struct sha1_context *context)
+{
+   if (!context)
+      return;
+
+   context->Length_Low             = 0;
+   context->Length_High            = 0;
+   context->Message_Block_Index    = 0;
+
+   context->Message_Digest[0]      = 0x67452301;
+   context->Message_Digest[1]      = 0xEFCDAB89;
+   context->Message_Digest[2]      = 0x98BADCFE;
+   context->Message_Digest[3]      = 0x10325476;
+   context->Message_Digest[4]      = 0xC3D2E1F0;
+
+   context->Computed   = 0;
+   context->Corrupted  = 0;
+}
+
+static void SHA1ProcessMessageBlock(struct sha1_context *context)
+{
+   const unsigned K[] =            /* Constants defined in SHA-1   */
+   {
+      0x5A827999,
+      0x6ED9EBA1,
+      0x8F1BBCDC,
+      0xCA62C1D6
+   };
+   int         t;                  /* Loop counter                 */
+   unsigned    temp;               /* Temporary word value         */
+   unsigned    W[80];              /* Word sequence                */
+   unsigned    A, B, C, D, E;      /* Word buffers                 */
+
+   /* Initialize the first 16 words in the array W */
+   for(t = 0; t < 16; t++)
+   {
+      W[t] = ((unsigned) context->Message_Block[t * 4]) << 24;
+      W[t] |= ((unsigned) context->Message_Block[t * 4 + 1]) << 16;
+      W[t] |= ((unsigned) context->Message_Block[t * 4 + 2]) << 8;
+      W[t] |= ((unsigned) context->Message_Block[t * 4 + 3]);
+   }
+
+   for(t = 16; t < 80; t++)
+      W[t] = SHA1CircularShift(1,W[t-3] ^ W[t-8] ^ W[t-14] ^ W[t-16]);
+
+   A = context->Message_Digest[0];
+   B = context->Message_Digest[1];
+   C = context->Message_Digest[2];
+   D = context->Message_Digest[3];
+   E = context->Message_Digest[4];
+
+   for(t = 0; t < 20; t++)
+   {
+      temp  =  SHA1CircularShift(5,A) +
+         ((B & C) | ((~B) & D)) + E + W[t] + K[0];
+      temp &= 0xFFFFFFFF;
+      E     = D;
+      D     = C;
+      C     = SHA1CircularShift(30,B);
+      B     = A;
+      A     = temp;
+   }
+
+   for(t = 20; t < 40; t++)
+   {
+      temp  = SHA1CircularShift(5,A) + (B ^ C ^ D) + E + W[t] + K[1];
+      temp &= 0xFFFFFFFF;
+      E     = D;
+      D     = C;
+      C     = SHA1CircularShift(30,B);
+      B     = A;
+      A     = temp;
+   }
+
+   for(t = 40; t < 60; t++)
+   {
+      temp  = SHA1CircularShift(5,A) +
+         ((B & C) | (B & D) | (C & D)) + E + W[t] + K[2];
+      temp &= 0xFFFFFFFF;
+      E     = D;
+      D     = C;
+      C     = SHA1CircularShift(30,B);
+      B     = A;
+      A     = temp;
+   }
+
+   for(t = 60; t < 80; t++)
+   {
+      temp = SHA1CircularShift(5,A) + (B ^ C ^ D) + E + W[t] + K[3];
+      temp &= 0xFFFFFFFF;
+      E = D;
+      D = C;
+      C = SHA1CircularShift(30,B);
+      B = A;
+      A = temp;
+   }
+
+   context->Message_Digest[0] =
+      (context->Message_Digest[0] + A) & 0xFFFFFFFF;
+   context->Message_Digest[1] =
+      (context->Message_Digest[1] + B) & 0xFFFFFFFF;
+   context->Message_Digest[2] =
+      (context->Message_Digest[2] + C) & 0xFFFFFFFF;
+   context->Message_Digest[3] =
+      (context->Message_Digest[3] + D) & 0xFFFFFFFF;
+   context->Message_Digest[4] =
+      (context->Message_Digest[4] + E) & 0xFFFFFFFF;
+
+   context->Message_Block_Index = 0;
+}
+
+static void SHA1PadMessage(struct sha1_context *context)
+{
+   if (!context)
+      return;
+
+   /*
+    *  Check to see if the current message block is too small to hold
+    *  the initial padding bits and length.  If so, we will pad the
+    *  block, process it, and then continue padding into a second
+    *  block.
+    */
+   context->Message_Block[context->Message_Block_Index++] = 0x80;
+
+   if (context->Message_Block_Index > 55)
+   {
+      while(context->Message_Block_Index < 64)
+         context->Message_Block[context->Message_Block_Index++] = 0;
+
+      SHA1ProcessMessageBlock(context);
+   }
+
+   while(context->Message_Block_Index < 56)
+      context->Message_Block[context->Message_Block_Index++] = 0;
+
+   /*  Store the message length as the last 8 octets */
+   context->Message_Block[56] = (context->Length_High >> 24) & 0xFF;
+   context->Message_Block[57] = (context->Length_High >> 16) & 0xFF;
+   context->Message_Block[58] = (context->Length_High >> 8) & 0xFF;
+   context->Message_Block[59] = (context->Length_High) & 0xFF;
+   context->Message_Block[60] = (context->Length_Low >> 24) & 0xFF;
+   context->Message_Block[61] = (context->Length_Low >> 16) & 0xFF;
+   context->Message_Block[62] = (context->Length_Low >> 8) & 0xFF;
+   context->Message_Block[63] = (context->Length_Low) & 0xFF;
+
+   SHA1ProcessMessageBlock(context);
+}
+
+static int SHA1Result(struct sha1_context *context)
+{
+   if (context->Corrupted)
+      return 0;
+
+   if (!context->Computed)
+   {
+      SHA1PadMessage(context);
+      context->Computed = 1;
+   }
+
+   return 1;
+}
+
+static void SHA1Input(struct sha1_context *context,
+      const unsigned char *message_array,
+      unsigned length)
+{
+   if (!length)
+      return;
+
+   if (context->Computed || context->Corrupted)
+   {
+      context->Corrupted = 1;
+      return;
+   }
+
+   while(length-- && !context->Corrupted)
+   {
+      context->Message_Block[context->Message_Block_Index++] =
+         (*message_array & 0xFF);
+
+      context->Length_Low += 8;
+      /* Force it to 32 bits */
+      context->Length_Low &= 0xFFFFFFFF;
+      if (context->Length_Low == 0)
+      {
+         context->Length_High++;
+         /* Force it to 32 bits */
+         context->Length_High &= 0xFFFFFFFF;
+         if (context->Length_High == 0)
+            context->Corrupted = 1; /* Message is too long */
+      }
+
+      if (context->Message_Block_Index == 64)
+         SHA1ProcessMessageBlock(context);
+
+      message_array++;
+   }
+}
+
+static int sha1_calculate(const char *path, char *result)
+{
+   struct sha1_context sha;
+   unsigned char buff[4096];
+   int rv    = 1;
+   RFILE *fd = filestream_open(path,
+         RETRO_VFS_FILE_ACCESS_READ,
+         RETRO_VFS_FILE_ACCESS_HINT_NONE);
+
+   if (!fd)
+      goto error;
+
+   buff[0] = '\0';
+
+   SHA1Reset(&sha);
+
+   do
+   {
+      rv = (int)filestream_read(fd, buff, 4096);
+      if (rv < 0)
+         goto error;
+
+      SHA1Input(&sha, buff, rv);
+   }while(rv);
+
+   if (!SHA1Result(&sha))
+      goto error;
+
+   sprintf(result, "%08X%08X%08X%08X%08X",
+         sha.Message_Digest[0],
+         sha.Message_Digest[1],
+         sha.Message_Digest[2],
+         sha.Message_Digest[3], sha.Message_Digest[4]);
+
+   filestream_close(fd);
+   return 0;
+
+error:
+   if (fd)
+      filestream_close(fd);
+   return -1;
+}
+#endif  // __LIBRETRO__
 
 #if defined(_MSC_VER)
 struct CALCSHA1 {
